requirements.txt
Adafruit-Blinka==8.23.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-fingerprint==2.2.14
adafruit-circuitpython-requests==2.0.2
adafruit-circuitpython-typing==1.9.5
Adafruit-PlatformDetect==3.53.0
Adafruit-PureIO==1.1.11
engineering-notation==0.10.0
fingerprint-enhancer==0.0.13
fingerprint-feature-extractor==0.0.10
imageio==2.31.6
lazy_loader==0.3
networkx==3.2.1
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.0.1
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
scikit-image==0.22.0
scipy==1.11.3
tifffile==2023.9.26
tk==0.1.0
typing_extensions==4.8.0
